Fix getting video orientation & duration for android <29

+++ android/src/main/kotlin/com/kasem/media_picker_builder/model/MediaAsset.kt
@@ -6,7 +6,7 @@ data class MediaAsset(
         val id: Long,
         val dateAdded: Long,
         val orientation: Int,
-        val duration: Long?,
+        val duration: Double?,
         val type: MediaFile.MediaType,
         val isLivePhoto: Boolean
 ) {
